Automatic Update Feature (#424)
* Update advanced.php

Added a checkbox in the Advanced Settings page to enable Automatic Update

* Update install_config.sh

The AUTOMATIC_UPDATE parameter needs to be created in the BirdNET-Pi configuration file.

* Update install_services.sh

Add a cron job to the system to performa automatic updates.

* Create automatic_update.cron

The line to be added to the cron table to check for need to perform automatic update.

* Create automatic_update.sh

Script to perform the automatic update if the flag is set in the configuration.  Also, the services are restarted after the update.

* Update install_services.sh

* Update automatic_update.sh

* Update advanced.php

Correct typo

* Update update_birdnet_snippets.sh

Add the automatic_update.cron to the list of cron jobs

* Update update_birdnet_snippets.sh

* Update automatic_update.cron

Change automatic update to once a week on Sunday at 3AM

* Update update_birdnet_snippets.sh

Adjusted to add option to existing installations

* Update automatic_update.sh

Removed unnecessary call to restart_services.sh

* Update advanced.php

Cleaned up the Options section line spacing in the settings table

* Update templates/automatic_update.cron

Co-authored-by: Nachtzuster <Nachtzuster@users.noreply.github.com>

* Update scripts/update_birdnet_snippets.sh

Co-authored-by: Nachtzuster <Nachtzuster@users.noreply.github.com>

* Delete scripts/automatic_update.sh

* Update advanced.php

Modified the description of the Automatic Update feature on the Advanced Configuration tab.

* Update install_config.sh

---------

Co-authored-by: Nachtzuster <Nachtzuster@users.noreply.github.com>